+++ fj-doc-base-kotlin/src/main/java/org/fugerit/java/doc/base/kotlin/dsl/Background.kt
@@ -1,0 +1,9 @@
+package org.fugerit.java.doc.base.kotlin.dsl
+
+class Background : HelperDSL.TagWithText( "background" ) {
+   fun image( init: Image.() -> Unit = {} ): Image {
+       return initTag(Image(), init);
+   }
+
+
+}