+++ fj-doc-base-kotlin/src/main/java/org/fugerit/java/doc/base/kotlin/dsl/Nbsp.kt
@@ -1,0 +1,7 @@
+package org.fugerit.java.doc.base.kotlin.dsl
+
+class Nbsp : HelperDSL.TagWithText( "nbsp" ) {
+
+   fun length( value: Int ): Nbsp = lengthType( this, "length", value )
+
+}