+++ fj-doc-base-kotlin/src/main/java/org/fugerit/java/doc/base/kotlin/dsl/PageBreak.kt
@@ -1,0 +1,6 @@
+package org.fugerit.java.doc.base.kotlin.dsl
+
+class PageBreak : HelperDSL.TagWithText( "page-break" ) {
+
+
+}