+++ fj-doc-base/src/main/java/org/fugerit/java/doc/base/facade/DocFacadeSource.java
@@ -38,7 +38,9 @@ public class DocFacadeSource {
 	public static final int SOURCE_TYPE_YAML = 3;
 	
 	public static final int SOURCE_TYPE_JSON_NG = 4;
-	
+
+	public static final int SOURCE_TYPE_KOTLIN = 9;
+
 	public static final int SOURCE_TYPE_DEFAULT = SOURCE_TYPE_XML;
 	
 	private static final Logger logger = LoggerFactory.getLogger( DocFacadeSource.class );
@@ -56,13 +58,16 @@ public class DocFacadeSource {
 	private static final String TYPE_SOURCE_YAML = "org.fugerit.java.doc.yaml.parse.DocYamlParser";		// yaml parser may not be in class loader
 	
 	private static final String TYPE_SOURCE_JSON_NG = "org.fugerit.java.doc.json.ng.DocJsonParserNG";   // json parser NG may not be in class loader
-	
+
+	private static final String TYPE_SOURCE_KOTLIN = "org.fugerit.java.doc.base.kotlin.parse.DocKotlinParser";	// kotlin parser may not be in class loader
+
 	private static final Properties PARSERS = new Properties();
 	static {
 		PARSERS.setProperty( String.valueOf( SOURCE_TYPE_XML ) , TYPE_SOURCE_XML );
 		PARSERS.setProperty( String.valueOf( SOURCE_TYPE_JSON ) , TYPE_SOURCE_JSON );
 		PARSERS.setProperty( String.valueOf( SOURCE_TYPE_YAML ) , TYPE_SOURCE_YAML );
 		PARSERS.setProperty( String.valueOf( SOURCE_TYPE_JSON_NG ) , TYPE_SOURCE_JSON_NG );
+		PARSERS.setProperty( String.valueOf( SOURCE_TYPE_KOTLIN ) , TYPE_SOURCE_KOTLIN );
 	}
 	
 	public DocParser getParserForSource( int sourceType ) {